feat : 코드 컨벤션에 맞춘 개행 수정, 주문 엔티티 필드 totalPrice 타입 변경

+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/entity/Order.java
@@ -20,7 +20,7 @@ public class Order {
     private Integer quantity;
 
     @Column(nullable = false)
-    private Long totalPrice;
+    private Integer totalPrice;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
@@ -45,7 +45,7 @@ public class Order {
     protected Order() {
     }
 
-    public Order(int quantity, long totalPrice, DeliveryStatus deliveryStatus, Customer customer, Manager manager, Product product) {
+    public Order(int quantity, int totalPrice, DeliveryStatus deliveryStatus, Customer customer, Manager manager, Product product) {
         this.quantity = quantity;
         this.totalPrice = totalPrice;
         this.deliveryStatus = deliveryStatus;
@@ -66,7 +66,7 @@ public class Order {
         return quantity;
     }
 
-    public long getTotalPrice() {
+    public int getTotalPrice() {
         return totalPrice;
     }
